// pnpm_packages/web/src/sdk/models/sessionData.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { SessionDataTeamId } from './sessionDataTeamId';
import type { SessionDataWorkflowId } from './sessionDataWorkflowId';
import type { SessionDataAgentData } from './sessionDataAgentData';
import type { SessionDataTeamData } from './sessionDataTeamData';
import type { SessionDataWorkflowData } from './sessionDataWorkflowData';
import type { SessionDataMetadata } from './sessionDataMetadata';
import type { SessionDataSummary } from './sessionDataSummary';

export interface SessionData {
  session_id: string;
  session_type: string;
  agent_id: string;
  team_id: SessionDataTeamId;
  workflow_id: SessionDataWorkflowId;
  user_id: string;
  agent_data: SessionDataAgentData;
  team_data: SessionDataTeamData;
  workflow_data: SessionDataWorkflowData;
  metadata: SessionDataMetadata;
  summary: SessionDataSummary;
  created_at: number;
  updated_at: number;
}